various fixes (#60)
PR includes bug fix where fastq would remove the last character of the last read's quality when at EOF.

This PR mostly just improved the typing for external functions and added a little bit to the sequencing package. The original PR #60 contained indexing for fastq files, but this should be added when it is actually needed. The implementation, however, did work.