better barcoding (#74)
This commit/PR improves on an edge case in barcoding where the sequencing read has both barcodes, but one is hanging off the edge (due to the nanopore kinda falling off at the last ~10bp). We do this by also considering the primer used to amplify the dual barcode, and taking the score from there. Included is a regression test, as well as all the default dual barcodes that I personally use.