add genbank parser to python (#87)
This adds a genbank parser to python through the C shared library. Most of this was written by AI, because it's pretty annoying to write, and is mostly just piping data. Amazingly, it seems to work. I also added a little benchmark showing that we have comparable performance to biopython - but I'm sure some of our other features will be even better in that regard.

- Keoni